Downgrade libraries that require compileSDKversion 31
Some google librararies that we use in Widget SDK require clients to have compileSDKVersion to be 31+. And even though devs can override the library version it was agreen to downgrade those libraries to latest version that would work with compileSDKVersion 30.

Changed the test app compile SDK version to 30 to make sure that no breaking changes would be introduced.

MUIC-721