Add workflow to increment Core SDK version
Using the previous work done in Gradle, a new workflow has been added
to `bitrise.yml` that increments the Core SDK version. A pull request
is created also by directly calling the GitHub API. This will be used
in the Core SDK release process, in order to increment the Core SDK
version used in the widgets once the Core SDK is released. This
workflow requires an environment variable called `NEW_VERSION`.

MOB-1510