Add workflows to automate version increases
The workflow `increment_project_version` increases the project version
after a release has been done. The same approach that is used for the
Core SDK is used here. The workflow `upgrade_dependencies` upgrades
the Core SDK version used in this project. This workflow is going to
be called from the Core SDK workflows.

MOB-1510