Update Bitrise deployment workflow
Instead of going to GitHub and manually generating a release, the
deployment workflow will be in charge of generating the release, the
tag, and the changelog. The only thing needed from our side is to send
it the NEW_VERSION environment variable. To make this even easier, a
new GitHub Action is created to do exactly that. Just send the version
through GitHub’s interface, and it will do everything. The tag trigger
is not available anymore.

Apart from that, the increment_project_version is updated so that it
doesn’t have unnecessary steps. This time it will succeed always.

MOB-1589