Add post release workflow
Because our android-sdk-widgets release depends on Maven Central, then
we cannot increase the android-sdk-widgets dependency in
android_cortex_banking_app until Maven Central reflects our release.
This step thus needs to be done manually. The tag to be sent is the
same as the released version.

I also had to re-add the spaces that were deleted in a previous PR
because this absolutely screws up the formatting inside Bitrise and
makes all steps with scripts to fail because commands show up
overlapped.

MOB-1589