Fix increment project version GitHub Action not using sent parameter
Bitrise does not override an envvar sent through URL, so the GitHub
Action doesn't take the parameter sent, so it always does a `patch`
version increment. With this change (suggested by Bitrise itself),
GitHub sends a `GITHUB_VERSION_INCREMENT_TYPE` which then updates
the actual `VERSION_INCREMENT_TYPE` in Bitrise. It's a convoluted
approach but there's no other way of doing it.

MOB-1610