Fix incorrect changelogs being generated on release
The `git-clone` step was fetching the repository without tags. This
made the changelog step not know what is the latest tag, thus
generating a changelog with all commits in the repo. By fetching the
tags, Bitrise will be able to generate a changelog from the latest tag
to the most recent commit.

MOB-1610