Add runtime theme support for CV media upgrade dialogs
Also updated a bit of code for all other dialogs as some of them missed font family in buttons and to overall make the code more consistent

Additionally, it seems that `BaseConfigurableButton.setTheme(UiTheme)` does not apply themes correctly. I marked it as deprecated and replaced in `Dialogs.kt`.

MOB-1991