Fix Activity onPause -> onResume breaking the permission dialog
The issue was caused by the DialogController addCallback() method, which always fired the current dialog state.

CallActivity and ChatActivity always add/remove the callback on Activity onResume/onPause to prevent the same dialog from appearing on both Activities. But as a result, because of this, a duplicate second dialog was being with Activity having no callbacks attached to it.

Ideally, need the full refactoring of the Alert dialogs logic in Widgets SDK. Asked this refactoring request to be added to the Android tech depts.

MOB-2171