Fix ActivityWatchers loosing refference to current Activity
ActivityWatchers are losing reference to current activity on screen rotation when multiple activities are being recreated and for some reason, activity from the back gets 'resumed' for a short moment (probably it was partially visible) and then again paused. As a result, ActivityWhatcher thought that all activities are paused.

MOB-2180
MOB-2179