Update paparazzi version
This change is related to an issue discovered in another PR:
https://github.com/salemove/android-sdk-widgets/pull/1125

For some reason, the previous version of Paparazzi had a problem applying View#isInEditMode, so some strings were not retrieved correctly.

I did it as a separate PR for easier review because all snapshots had very slight changes because of the library version change.