Fix SDK license compliance
Renamed `LICENSE` file to `LICENSE.txt` to fix link in old maven releases that have hardcoded link to license in POM file.

Added license file in the src/main/assets directory. By adding it there it will be included in the AAR’s assets folder. While placing the license at the root of the AAR might have made it more visible, this approach still ensures that the license is packaged with the library, easily accessible to consumers, and compliant with licensing requirements—all without the need for post-processing the AAR through custom Gradle scripts.

The main minus of this solution is that now there are two copies of license file in the same project making it easy to miss in later changes.

MOB-3835