Fix IllegalStateException when starting NotificationRemovalService
We where not able to reproduce the issue which was reported by the client. However, theoretically Lifecycle.Event.ON_STOP might be called when app is in bacground causing the error 'IllegalStateException: Not allowed to start service app is in background' but in case of Lifecycle.Event.ON_PAUSE app can't be in the background state.

MOB-4011